requirements.txt
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
colorama==0.4.6
confection==0.1.5
cryptography==43.0.0
cssselect==1.2.0
cymem==2.0.8
distro==1.8.0
dnspython==2.6.1
duckduckgo-search==3.0.2
en-core-web-md @ https://github.com/explosion/spacy-models/releases/download/en_core_web_md-3.7.1/en_core_web_md-3.7.1-py3-none-any.whl
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.2.2
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.19.1
google-api-python-client==2.141.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
groq==0.9.0
gTTS==2.3.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
langcodes==3.4.0
language_data==1.2.0
loguru==0.7.2
lxml==5.3.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
murmurhash==1.0.10
numpy==1.26.4
orjson==3.8.10
outcome==1.3.0.post0
packaging==24.1
pillow==10.4.0
pinecone-client==2.2.1
playsound==1.2.2
preshed==3.0.9
proto-plus==1.24.0
protobuf==5.27.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.18.0
pylatexenc==2.10
pyOpenSSL==24.2.1
pyparsing==3.1.2
PyPDF2==3.0.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
PyYAML==6.0
readability-lxml==0.8.1
redis==5.0.8
referencing==0.35.1
regex==2024.7.24
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
rsa==4.9
selenium==4.1.4
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
thinc==8.2.5
tqdm==4.66.5
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.4
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.19
urllib3-secure-extra==0.1.0
wasabi==1.1.3
weasel==0.4.1
webdriver-manager==4.0.2
wrapt==1.16.0
wsproto==1.2.0
